gohome

+++ objectdetection/src/detector.h
@@ -16,8 +16,8 @@ using namespace std;
 using namespace cv;
 
 struct RegressionLine {
-    int delta_x;
-    int delta_y; //x from -95 to 95 in px y from 0 to 240
+    float delta_x;
+    float delta_y; //x from -95 to 95 in px y from 0 to 240
     int intercept; //y from the middle
     float error;
 };
